+++ crates/store/src/driver.rs
@@ -375,10 +375,13 @@ impl StoreDriverOn {
                 self.store.storage_mut().disarm_interruption();
                 let model_result = self.model.apply(operation);
                 if store_result != model_result {
-                    return Err((self.store, StoreInvariant::DifferentResult {
-                        store: store_result,
-                        model: model_result,
-                    }));
+                    return Err((
+                        self.store,
+                        StoreInvariant::DifferentResult {
+                            store: store_result,
+                            model: model_result,
+                        },
+                    ));
                 }
                 if store_result.is_ok() {
                     if let Err(invariant) = self.check_deleted(&deleted) {