+++ scripts/wrapper.sh
@@ -49,11 +49,11 @@ case "$1" in
       *) e "Wrapper does not support 'cargo $2'" ;;
     esac
     ;;
-  mdbook) ensure_cargo mdbook 0.4.43 ;;
-  probe-rs) ensure_cargo probe-rs-tools 0.25.0 ;;
+  mdbook) ensure_cargo mdbook 0.4.45 ;;
+  probe-rs) ensure_cargo probe-rs-tools 0.27.0 ;;
   rust-objcopy|rust-size) ensure_cargo cargo-binutils 0.3.6 ;;
   taplo) ensure_cargo taplo-cli 0.9.3 ;;
-  trunk) ensure_cargo trunk 0.21.5 ;;
+  trunk) ensure_cargo trunk 0.21.7 ;;
   twiggy) ensure_cargo twiggy 0.7.0 ;;
   *) IS_CARGO=n ;;
 esac